tests/python/requirements-test.txt
jsonschema>=4.20
pyyaml>=6.0
pytest>=7.4
